// src/Products/interface.ts
export interface ProductState {
    products: IProduct[],
    order: 'asc' | 'desc'
}

export interface IProduct {
    id: number
    title: string
    price: number
    description: string
    category: string
    image: string
    rating: Rating
}

export interface Rating {
    rate: number
    count: number
}